refactor ChannelCard component
