+++ src/shared/components/ChannelCardBase/ChannelCardBase.stories.tsx
@@ -2,11 +2,11 @@ import { Meta, Story } from '@storybook/react'
 import React from 'react'
 import { BrowserRouter } from 'react-router-dom'
 
-import { TempChannelCardBase, TempChannelCardBaseProps } from './TempChannelCardBase'
+import { ChannelCardBase, ChannelCardBaseProps } from './ChannelCardBase'
 
 export default {
   title: 'Shared/C/ChannelCardBase',
-  component: TempChannelCardBase,
+  component: ChannelCardBase,
   args: {
     avatarUrl: 'https://eu-central-1.linodeobjects.com/atlas-assets/channel-avatars/2.jpg',
     title: 'Example Channel',
@@ -23,10 +23,10 @@ export default {
   ],
 } as Meta
 
-const Template: Story<TempChannelCardBaseProps> = (args) => {
+const Template: Story<ChannelCardBaseProps> = (args) => {
   return (
     <>
-      <TempChannelCardBase {...args} />
+      <ChannelCardBase {...args} />
     </>
   )
 }
