--- src/shared/components/TempChannelCardBase/index.ts
@@ -1,1 +1,0 @@
-export * from './TempChannelCardBase'
