+++ src/shared/components/index.ts
@@ -7,7 +7,6 @@ export * from './RadioButton'
 export * from './Text'
 export * from './VideoTileBase'
 export * from './VideoPlayer'
-export * from './ChannelCardBase'
 export * from './HamburgerButton'
 export * from './Gallery'
 export * from './ChannelAvatar'
@@ -47,4 +46,4 @@ export * from './EmptyFallback'
 export * from './LoadMoreButton'
 export * from './CallToActionButton'
 export * from './GridHeading'
-export * from './TempChannelCardBase'
+export * from './ChannelCardBase'
